+++ CAN_DBC_DECODE_TOOL/dbc_parser.hpp
@@ -23,8 +23,8 @@ public:
 	// A bool is used to indicate whether parsing succeeds or not
 	bool parse(const std::string& filePath);
 	// Decode
-	std::unordered_map<std::string, double> decode(uint32_t msgId, unsigned char payLoad[], unsigned short dlc);
-	double decodeSignalOnRequest(uint32_t msgId, unsigned char payLoad[], unsigned short dlc, std::string sigName);
+	std::unordered_map<std::string, double> decode(unsigned int msgId, unsigned char payLoad[], unsigned int dlc);
+	double decodeSignalOnRequest(unsigned int msgId, unsigned char payLoad[], unsigned int dlc, std::string sigName);
 	// Print DBC Info
 	void printDbcInfo();
 
@@ -37,7 +37,7 @@ private:
 	typedef std::vector<Message*> messages_t;
 	messages_t messagesInfo;
 	// A hash table that stores all info of messages. <Message id, Message object>
-	std::unordered_map<uint32_t, Message> messageLibrary;
+	std::unordered_map<unsigned int, Message> messageLibrary;
 	// Function used to parse DBC file
 	void loadAndParseFromFile(std::istream& in);
